make test04-jumpsret pass
Two issues here: first, JSR instructions in the trace disassembly
include what is I assume the name of the subroutine being called, which
I need to crop out when loading in the traces. Next, because of the
branches, the execution is nonlinear so the traced disassembly might
have fewer instructions than the ROM disassembly. The test case was
updated to account for this.